Navigate to ReportBaseView
ReportBaseView containing charts, reports
