+++ ViewModels/StaffViewModels/ManagerViewModel.cs
@@ -12,22 +12,7 @@ namespace HotelManager.ViewModels.StaffViewModels
 {
     public class ManagerViewModel : BaseViewModel
     {
-        public ObservableCollection<string> Options { get; set; }
 
-        private string _selectedOption;
-        public string SelectedOption
-        {
-            get => _selectedOption;
-            set
-            {
-                if (_selectedOption != value)
-                {
-                    _selectedOption = value;
-                    OnPropertyChanged(nameof(SelectedOption));
-                    changeContentControl(_selectedOption);
-                }
-            }
-        }
 
         private Object _currentContent;
         public Object CurrentContent
@@ -45,19 +30,16 @@ namespace HotelManager.ViewModels.StaffViewModels
 
         // commands
         public IRelayCommand ShowEmployeeListCommand { get; }
+        public IRelayCommand ShowReportsCommand { get; }
+
 
         public ManagerViewModel()
         {
-            Options = new ObservableCollection<string>
-            {
-                "View revenue report",
-                "View receptionist activity report"
-            };
-            // Set a default selected option
-            SelectedOption = Options.FirstOrDefault();
-
             // init command
             ShowEmployeeListCommand = new RelayCommand(ShowEmployeeList);
+            ShowReportsCommand = new RelayCommand(ShowReports);
+
+            ShowReports();
         }
 
         private void ShowEmployeeList()
@@ -71,28 +53,15 @@ namespace HotelManager.ViewModels.StaffViewModels
             CurrentContent = new Views.ManagerViews.EmployeesListView();
         }
 
-
-        void changeContentControl(string option)
+        private void ShowReports()
         {
-            // Dispose of the old view's DataContext if it implements IDisposable
-            // prevent memory leaks
+            // Dispose old content nếu có
             if (CurrentContent is FrameworkElement oldView && oldView.DataContext is IDisposable disposable)
             {
                 disposable.Dispose();
             }
 
-            switch (option)
-            {
-                case "View revenue report":
-                    CurrentContent = new Views.ManagerViews.RevenueReportChartView();
-                    break;
-                case "View receptionist activity report":
-                    CurrentContent = new Views.ManagerViews.ReceptionistActivityReportChartView();
-                    break;
-                default:
-                    CurrentContent = new Views.ManagerViews.RevenueReportChartView();
-                    break;
-            }
+            CurrentContent = new Views.ManagerViews.ReportsBaseView();
         }
     }
 }
