+++ Views/ManagerViews/EmployeesListView.xaml
@@ -5,7 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              
              xmlns:vm="clr-namespace:HotelManager.ViewModels.ManagerViewModels"
-                     xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
              
              mc:Ignorable="d" 
              d:DesignHeight="500" d:DesignWidth="800"
