+++ Views/ManagerViews/ReportsBaseView.xaml
@@ -1,0 +1,63 @@
+﻿<UserControl x:Class="HotelManager.Views.ManagerViews.ReportsBaseView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:HotelManager.Views.ManagerViews"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800"
+             
+             xmlns:vm="clr-namespace:HotelManager.ViewModels.ManagerViewModels"
+             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes">
+
+    <UserControl.DataContext>
+        <vm:ReportsBaseViewModel/>
+    </UserControl.DataContext>
+
+    <Grid>
+        
+        <materialDesign:Card Grid.Row="1" 
+                             Effect="{StaticResource Elevation3}"
+                             Background="{StaticResource SurfaceBrush}">
+            <ContentControl x:Name="CurrentContent"
+                            Content="{Binding CurrentContent}"
+                            Margin="0,0,0,0">
+
+                <ContentControl.Triggers>
+                    <EventTrigger RoutedEvent="Loaded">
+                        <BeginStoryboard>
+                            <Storyboard>
+                                <DoubleAnimation Storyboard.TargetProperty="Opacity"
+                                                 From="0" To="1" Duration="0:0:0.5"/>
+                                <ThicknessAnimation Storyboard.TargetProperty="Margin"
+                                                    From="40,40,40,40" To="24,24,24,24" 
+                                                    Duration="0:0:0.5"/>
+                            </Storyboard>
+                        </BeginStoryboard>
+                    </EventTrigger>
+                </ContentControl.Triggers>
+            </ContentControl>
+        </materialDesign:Card>
+
+        <ComboBox x:Name="comboBox_Options"
+                  Width="auto"
+                  Height="auto"
+                  Margin="20,20,20,20"
+                  HorizontalAlignment="Right"
+                  VerticalAlignment="Top"
+                  ItemsSource="{Binding Options}"
+                  SelectedItem="{Binding SelectedOption, Mode=TwoWay}"
+                  materialDesign:HintAssist.Hint="Select View"
+                  materialDesign:HintAssist.IsFloating="True"
+                  Style="{StaticResource MaterialDesignFilledComboBox}"
+                  Background="{StaticResource Error100Brush}"
+                  Foreground="Navy" >
+            <ComboBox.Resources>
+                <Style TargetType="ComboBoxItem">
+                    <Setter Property="Foreground" Value="{StaticResource OnSurfaceBrush}"/>
+                </Style>
+            </ComboBox.Resources>
+        </ComboBox>
+
+    </Grid>
+</UserControl>
